+++ src/tests/pages/FormPage.test.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { render } from "@testing-library/react";
 
 import { MemoryRouter } from "react-router-dom";
-import FormPage from "../../pages/FormPage";
+import FormPage from "../../pages/FormPage/FormPage";
 
 import * as forms from "../../api/forms";
